feat: fix bot upgrade contracts

+++ crates/pbh-helpers/src/bindings.rs
@@ -4,8 +4,8 @@ sol! {
     #[sol(rpc)]
     interface IPBHKotH {
         function ctf(address receiver) external;
-        function gameEnd() external view returns (uint128);
-        function latestBlock() external view returns (uint128);
+        function gameEnd() external view returns (uint256);
+        function latestBlock() external view returns (uint256);
         function leader() external view returns (address);
         function highScore() external view returns (uint256);
         function leaderboard(address addr) external view returns (uint256);
@@ -13,6 +13,6 @@ sol! {
 
     #[sol(rpc)]
     interface IPBHEntryPoint {
-        function numPbhPerMonth() external view returns (uint8);
+        function numPbhPerMonth() external view returns (uint16);
     }
 }
